requirementsAIFeedSummarizer.txt
python-dotenv
feedparser